backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1